refactor(layout): Header 테마 드롭다운을 ThemeDropdown 컴포넌트로 추출
데스크탑/모바일 테마 드롭다운 JSX 중복을 ThemeDropdown 컴포넌트로 추출하여 SonarCloud 중복 임계치 준수 및 유지보수성 향상.
